Added support for "arc projection" in the radial direction.
It is not enough to just use the radial vector length, since it will not
be able to differentiate between a point and its mirror 180 degrees. This
was solved by using the same signa as the radial dot product that was used
before. Arc projection is now enabled by default. A flag that can be
configured through set_parameter() will be added later.
GCC compatibility: using std::copysign() instead if _copysign().